float to int

+++ 10labArray5Sum.cpp
@@ -14,7 +14,7 @@ using namespace std;
 int main()
 {
     //declare array values (hardcoded) and sum initial value
-    float array[5] = {1, 2, 3, 4, 5};
+    int array[5] = {1, 2, 3, 4, 5};
     float sum = 0;
 
     //declare i as 0; when i is less than 5, the number of values array holds, the statement is true and the code will run; adds 1 value to i after each loop
